requirements.txt
Django==4.2
djangorestframework==3.14.0
Faker==18.6.1
